refactor: move common tools for votableItems to VotableItemsService
